requirements.txt
asn1crypto==0.24.0
backports.csv==1.0.7
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.8.2
Click==7.0
cryptography==2.1.4
defusedxml==0.6.0
Django==1.11.29
django-bootstrap-form==3.4
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
django-filter==2.0.0
django-tables2==2.3.1
enum34==1.1.6
et-xmlfile==1.0.1
Flask==1.1.1
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
idna==2.6
ipaddress==1.0.17
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
keyring==10.6.0
keyrings.alt==3.0
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==1.4.4
numpy==1.16.5
odfpy==1.4.1
openpyxl==2.6.4
PyAudio==0.2.11
pycrypto==2.6.1
pygobject==3.26.1
PyMySQL==0.9.3
pytz==2019.3
pyxdg==0.25
PyYAML==5.3.1
SecretStorage==2.3.1
selenium==3.141.0
six==1.11.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
tablib==0.14.0
urllib3==1.25.6
virtualenv==16.7.5
Werkzeug==0.16.0
xlrd==1.2.0
xlwt==1.3.0
